chore: update create-plugin to 7.7.0
