+++ .config/.cprc.json
@@ -1,3 +1,3 @@
 {
-  "version": "5.25.8"
+  "version": "5.27.1"
 }
