+++ package.json
@@ -3,12 +3,12 @@
   "version": "2.1.6",
   "description": "Grafana IoT TwinMaker App Plugin",
   "scripts": {
-    "build": "webpack -c ./webpack.config.ts --env production",
+    "build": "webpack -c ./.config/webpack/webpack.config.ts --env production",
     "clean": "rm -rf dist/ build/*",
-    "dev": "webpack -w -c ./webpack.config.ts --env development",
+    "dev": "webpack -w -c ./.config/webpack/webpack.config.ts --env development",
     "e2e": "playwright test",
     "generate-release-notes": "PREV_TAG=$(git tag | tail -n 1) && gh api --method POST /repos/grafana/grafana-iot-twinmaker-app/releases/generate-notes -f tag_name=v${npm_package_version} -f target_commitish=main -f previous_tag_name=${PREV_TAG} | jq -r .body",
-    "lint": "eslint --cache .",
+    "lint": "eslint --cache --ignore-path ./.gitignore --ext .js,.jsx,.ts,.tsx .",
     "lint:fix": "yarn run lint --fix && prettier --write --list-different .",
     "server": "docker compose up --build",
     "sign": "npx --yes @grafana/sign-plugin@latest",
@@ -96,12 +96,12 @@
     "@cloudscape-design/global-styles": "^1",
     "@emotion/css": "11.13.5",
     "@grafana/aws-sdk": "0.8.3",
-    "@grafana/data": "^12.1.0",
+    "@grafana/data": "^12.2.0",
     "@grafana/plugin-ui": "^0.12.0",
-    "@grafana/runtime": "^12.1.0",
+    "@grafana/runtime": "^12.2.0",
     "@grafana/scenes": "^6.42.2",
-    "@grafana/schema": "^12.1.0",
-    "@grafana/ui": "^12.1.0",
+    "@grafana/schema": "^12.2.0",
+    "@grafana/ui": "^12.2.0",
     "@iot-app-kit/core": "10.13.1",
     "@iot-app-kit/react-components": "10.13.1",
     "@iot-app-kit/scene-composer": "10.13.1",
@@ -116,7 +116,8 @@
     "rxjs": "7.8.2",
     "semver": "^7.7.2",
     "tslib": "2.8.1",
-    "uuid": "^11.1.0"
+    "uuid": "^11.1.0",
+    "@grafana/i18n": "^12.2.0"
   },
   "engines": {
     "node": ">=24"
